@@ -1,7 +1,7 @@
 package com.hbelmiro.microprofile.cdi;
 
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
 
 @ApplicationPath("/portfolio")
 public class MicroprofileCdiRestApplication extends Application {
